Whois typo fixes
git-svn-id: http://svn.inspircd.org/repository/trunk/inspircd@3525 e03df62e-2008-0410-955e-edbf42e46eb7

+++ src/helperfuncs.cpp
@@ -1494,7 +1494,7 @@ bool DirValid(char* dirandfile)
         if( getcwd( otherdir, MAXBUF ) == NULL )
                 return false;
         chdir(buffer);
-	int t = strlen(work);
+	size_t t = strlen(work);
         if (strlen(otherdir) >= t)
         {
                 otherdir[t] = '\0';
